fix: fix contract overview screen
