+++ lib/core/routers/routers.dart
@@ -1,11 +1,8 @@
 import 'package:defifundr_mobile/core/routers/routes_constant.dart';
-import 'package:defifundr_mobile/screens/auth_screens/verify_bvn/verify_bvn_screen.dart';
-import 'package:defifundr_mobile/screens/auth_screens/verify_with_passport_screen/verify_with_passport.dart';
-import 'package:defifundr_mobile/screens/voters_verification_screen/votersVerification.dart';
+import 'package:defifundr_mobile/screens/auth_screens/contract_overview/view/contract_overview.dart';
 import 'package:defifundr_mobile/screens/auth_screens/create_password/create_password_screen.dart';
 import 'package:defifundr_mobile/screens/auth_screens/get_started/view/get_started.dart';
-import 'package:defifundr_mobile/screens/home/contract/presentation/view/compliance-page.dart';
-import 'package:defifundr_mobile/screens/home/identity/verify_with_drivers_card.dart';
+import 'package:defifundr_mobile/screens/home/contract/presentation/view/compliance_page.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
@@ -88,7 +85,7 @@ class AppRouter {
           pageBuilder: (context, state) {
             return CustomTransitionPage(
               key: state.pageKey,
-              child: const ComplianceScreen(),
+              child: const ContractOverview(),
               transitionsBuilder:
                   (context, animation, secondaryAnimation, child) {
                 return FadeTransition(
